AI/SilverLink-AI/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "LLM 기반 콜봇·챗봇 통합 지능형 상담 AI 시스템 API"
authors = [
    {name = "SeongHo Kim",email = "kkb053404@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "dependency-injector (>=4.48.3,<5.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "ruff (>=0.14.13,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "twilio (>=9.9.1,<10.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "openai (>=2.15.0,<3.0.0)",
    "psutil (>=7.2.1,<8.0.0)",
    "pymilvus (>=2.6.6,<3.0.0)",
    "langchain-core (>=1.2.7,<2.0.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "langgraph (>=1.0.7,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "faker (>=40.1.2,<41.0.0)",
    "boto3 (>=1.42.37,<2.0.0)",
    "huggingface-hub (>=1.3.5,<2.0.0)",
    "spacy (>=3.8.11,<4.0.0)",
    "presidio-analyzer (>=2.2.360,<3.0.0)",
    "presidio-anonymizer (>=2.2.360,<3.0.0)",
    "mem0ai (>=1.0.2,<2.0.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "rapidfuzz (>=3.0.0,<4.0.0)",
    "httpx (>=0.28.0,<1.0.0)",
    "chromadb (>=0.4.22,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"